支持了研究生学号

+++ Entity/Volunteer.cs
@@ -12,7 +12,7 @@ namespace VolunteerDatabase.Entity
     public class Volunteer:IComparable<Volunteer>, IEqualityComparer<Volunteer>
     {
         //[Required]
-        public int StudentNum { get; set;}//学号
+        public long StudentNum { get; set;}//学号
 
         [DatabaseGenerated(DatabaseGeneratedOption.Identity)]
         public Guid UID { get; set; }//唯一标识符：主键
